+++ editor/vscode/browser-language-server/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "slint-vscode-browser-language-server",
+    "description": "The Slint LSP worker for the VScode extension.",
+    "version": "0.2.5",
+    "author": "Slint",
+    "license": "GPL-3.0",
+    "engines": {
+        "node": "*"
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/slint-ui/slint"
+    },
+    "dependencies": {
+        "vscode-languageserver": "^7.0.0",
+        "vscode-languageserver-textdocument": "^1.0.1"
+    },
+    "scripts": {}
+}
