+++ tools/lsp/Cargo.toml
@@ -22,6 +22,11 @@ icon = ["../../logo/slint-logo-small-light-128x128.png"]
 name = "slint-lsp"
 path = "main.rs"
 
+[lib]
+# lib is there only for the web
+crate-type = ["cdylib"]
+path = "wasm_main.rs"
+
 [features]
 backend-qt = ["slint-interpreter/backend-qt"]
 backend-gl-all = ["slint-interpreter/backend-gl-all"]
@@ -46,3 +51,9 @@ once_cell = "1.9.0"
 serde = "1.0.118"
 serde_json = "1.0.60"
 spin_on = "0.1"
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen = { version = "0.2", features = ["serde-serialize"] }
+
+# [package.metadata.wasm-pack.profile.release]
+# wasm-opt = false
